Fix ambiguous implicit casts that stopped the Windows build from working - hopefully the last one, as I've checked all other CPPUNIT_ASSERT_EQUAL macro invocations

+++ tests/RDFTest.cpp
@@ -331,7 +331,7 @@ public:
          i++)
       (*i).second->release_ref();
 
-    CPPUNIT_ASSERT_EQUAL(0, mErrorCount);
+    CPPUNIT_ASSERT_EQUAL(static_cast<int32_t>(0), mErrorCount);
   }
 private:
   std::string
